server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.7.4
pillow==10.3.0
opencv-python==4.10.0.84
python-dotenv==1.0.1
numpy==1.26.4
requests==2.32.3
